requirements.txt
pytest==8.3.5
fastapi==0.115.12
pydantic==2.11.5
mysql-connector-python==9.3.0
streamlit==1.45.1
requests==2.32.3
pandas==2.2.3
uvicorn==0.34.2
contextlib==

#when someone using the project they can intall all the packages using the below command
#pip install -r .\requirements.txt
#this will install all the packages